Lesson 17. Adding checkboxes for labels into editing product form. Using mutators. Displaying labels in product card: New, Hit, Recommend# Please enter the commit message for your changes. Lines starting

+++ database/migrations/2020_09_16_093942_create_products_table.php
@@ -20,7 +20,7 @@ class CreateProductsTable extends Migration
             $table->string('code');
             $table->string('description')->nullable();
             $table->string('image')->nullable();
-            $table->double('price')->default(0);
+            $table->bigInteger('price')->default(0);
 
             $table->timestamps();
         });
